Update server link in workflow file
Realistically this needs to be either downloaded from within the container at runtime or updated automatically. I've asked the R5R team to give us a perma link to the latest build just like GitHub has with its /latest tag on published releases. This should solve the issue of the dedi link needed to be updated in this workflow every time